Code/RGT2/requirements.txt
jupyter
numpy
torch
hydra-core
transformers
scikit-learn
omegaconf
pronouncing

